backend/requirements.txt
# Core FastAPI and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database and ORM
asyncpg>=0.28.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Task queue and caching
redis>=4.5.0

# Search engine
elasticsearch>=8.0.0

# Authentication and security
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
cryptography>=40.0.0

# Data validation and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client
httpx>=0.24.0
aiofiles>=23.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
